Add a paymentrequests page
