+++ webapp/src/constants/configs.ts
@@ -15,6 +15,7 @@ export const MASTERNODE_LIST_PAGE_SIZE = 10;
 export const WALLET_TXN_PAGE_SIZE = 5;
 export const MAX_WALLET_TXN_PAGE_SIZE = 200;
 export const PAYMENT_REQ_PAGE_SIZE = 5;
+export const PAYMENT_REQ_LIST_SIZE = 10;
 export const QUEUE_CONCURRENCY = 5;
 export const BALANCE_CRON_DELAY_TIME = 300000;
 export const BITCOIN_CLI_REGEX = /bitcoin-cli/g;
