+++ webapp/src/constants/routespath.ts
@@ -14,6 +14,7 @@ export const WALLET_PAGE_PATH = '/';
 export const WALLET_BASE_PATH = '/wallet';
 export const WALLET_SEND_PATH = `${WALLET_BASE_PATH}/send`;
 export const WALLET_RECEIVE_PATH = `${WALLET_BASE_PATH}/receive`;
+export const WALLET_CREATE_RECEIVE_REQUEST = `${WALLET_BASE_PATH}/receive/request`;
 export const WALLET_PAYMENT_REQ_BASE_PATH = `${WALLET_BASE_PATH}/paymentrequest`;
 export const WALLET_PAYMENT_REQ_PARAMS_PATH = `${WALLET_PAYMENT_REQ_BASE_PATH}/:id`;
 export const CONSOLE_RPC_CALL_BASE_PATH = `/console`;
