+++ webapp/src/containers/WalletPage/components/ReceivePage/PaymentRequestList/index.tsx
@@ -1,0 +1,184 @@
+import React, { useEffect, useState } from 'react';
+import moment from 'moment';
+import {
+  Card,
+  CardBody,
+  Table,
+  UncontrolledDropdown,
+  DropdownToggle,
+  DropdownMenu,
+  DropdownItem,
+} from 'reactstrap';
+import { connect } from 'react-redux';
+import { I18n } from 'react-redux-i18n';
+import { MdMoreHoriz, MdDelete } from 'react-icons/md';
+import styles from './PaymentRequestList.module.scss';
+import {
+  fetchPaymentRequest,
+  removeReceiveTxnsRequest,
+} from '../../../reducer';
+import {
+  WALLET_PAYMENT_REQ_BASE_PATH,
+  PAYMENT_REQ_LIST_SIZE,
+} from '../../../../../constants';
+import QrCode from '../../../../../components/QrCode';
+import CopyToClipboard from '../../../../../components/CopyToClipboard';
+import Pagination from '../../../../../components/Pagination';
+import { getTransactionURI } from '../../../../../utils/utility';
+
+interface PaymentRequestsProps {
+  unit: string;
+  paymentRequests: {
+    label: string;
+    id: number;
+    time: string;
+    address: string;
+    message: string;
+    amount: number;
+    unit: string;
+  }[];
+  fetchPaymentRequest: () => void;
+  removeReceiveTxns: (id: string | number) => void;
+}
+
+const PaymentRequests: React.FunctionComponent<PaymentRequestsProps> = (
+  props: PaymentRequestsProps
+) => {
+  useEffect(() => {
+    props.fetchPaymentRequest();
+  }, []);
+
+  const [currentPage, handlePageClick] = useState(1);
+  const pageSize = PAYMENT_REQ_LIST_SIZE;
+  const total = props.paymentRequests.length;
+  const pagesCount = Math.ceil(total / pageSize);
+  const from = (currentPage - 1) * pageSize;
+  const to = Math.min(total, currentPage * pageSize);
+  const data = props.paymentRequests.slice(from, to);
+  const [copied, changeCopied] = useState(false);
+
+  const handleCopy = () => {
+    changeCopied(true);
+    setTimeout(() => {
+      changeCopied(false);
+    }, 600);
+  };
+  return (
+    <section className='mb-5'>
+      {total ? (
+        <>
+          <Card className='table-responsive-md'>
+            <Table className={styles.table}>
+              <thead>
+                <tr>
+                  <th></th>
+                  <th>
+                    {I18n.t('containers.wallet.paymentRequestList.label')}
+                  </th>
+                  <th>
+                    {I18n.t('containers.wallet.paymentRequestList.address')}
+                  </th>
+                  <th>
+                    {I18n.t('containers.wallet.paymentRequestList.created')}
+                  </th>
+                  <th></th>
+                  <th></th>
+                </tr>
+              </thead>
+              <tbody>
+                {data.map((request) => {
+                  const transactionURI = getTransactionURI(
+                    request.unit,
+                    request.address,
+                    {
+                      label: request.label,
+                      amount: request.amount,
+                      message: request.message,
+                    }
+                  );
+                  return (
+                    <tr key={request.id}>
+                      <td></td>
+                      <td>{request.label}</td>
+                      <td>{request.address}</td>
+                      <td>{moment(request.time).fromNow()}</td>
+                      <td>
+                        <div>
+                          <span>
+                            <QrCode
+                              value={transactionURI}
+                              uid={`${request.id}`}
+                              qrClass={styles.qr}
+                            />
+                          </span>
+                          <span>
+                            <CopyToClipboard
+                              value={transactionURI}
+                              handleCopy={handleCopy}
+                            />
+                          </span>
+                        </div>
+                      </td>
+                      <td className={styles.actionCell}>
+                        <UncontrolledDropdown>
+                          <DropdownToggle className='padless' color='link'>
+                            <MdMoreHoriz />
+                          </DropdownToggle>
+                          <DropdownMenu right>
+                            <DropdownItem
+                              onClick={() =>
+                                props.removeReceiveTxns(request.id)
+                              }
+                            >
+                              <MdDelete />
+                              <span>
+                                {I18n.t(
+                                  'containers.wallet.paymentRequests.cancelRequest'
+                                )}
+                              </span>
+                            </DropdownItem>
+                          </DropdownMenu>
+                        </UncontrolledDropdown>
+                      </td>
+                    </tr>
+                  );
+                })}
+              </tbody>
+            </Table>
+          </Card>
+          <Pagination
+            label={I18n.t('containers.wallet.walletPage.paginationRange', {
+              to,
+              total,
+              from: from + 1,
+            })}
+            currentPage={currentPage}
+            pagesCount={pagesCount}
+            handlePageClick={handlePageClick}
+          />
+        </>
+      ) : (
+        <Card className='table-responsive-md'>
+          <CardBody>
+            {I18n.t('containers.wallet.walletPage.noPaymentRequests')}
+          </CardBody>
+        </Card>
+      )}
+    </section>
+  );
+};
+
+const mapStateToProps = (state) => {
+  const { wallet, settings } = state;
+  return {
+    unit: settings.appConfig.unit,
+    paymentRequests: wallet.paymentRequests,
+  };
+};
+
+const mapDispatchToProps = {
+  fetchPaymentRequest,
+  removeReceiveTxns: (id: string | number) => removeReceiveTxnsRequest(id),
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(PaymentRequests);
